moved smartdocuments folder to new parameters-components folder
