+++ src/main/app/src/app/admin/parameters-components/smart-documents-form/smart-documents-form-item/smart-documents-form-item.component.less
