+++ src/main/app/src/app/admin/parameters-components/smart-documents-form/smart-documents-form.component.html
