+++ src/main/app/src/app/admin/parameters-edit-cmmn/parameters-edit-cmmn.component.ts
@@ -63,7 +63,7 @@ import {
 } from "../model/parameters/process-model-method";
 import { ReferentieTabelService } from "../referentie-tabel.service";
 import { ZaakafhandelParametersService } from "../zaakafhandel-parameters.service";
-import { SmartDocumentsFormComponent } from "./smart-documents-form/smart-documents-form.component";
+import { SmartDocumentsFormComponent } from "../parameters-components/smart-documents-form/smart-documents-form.component";
 
 /**
  * Form-local variant of RestZaakbeeindigParameter where zaakbeeindigReden and resultaattype
